Hopefully fix the "3MHz MCLKO2" bug for good
After another "smoke" event from the step-up MOSFET trying to desolder
itself and more investigations, it seems that the ACGDCTL register is
somewhat flaky to load.

My theory is that maybe it depends on having the right MLCKI2 clock edge
when loading...

Just writing it repeatedly seems to make loading more reliable.
For good measure, also load it on every SOF in the DPLL handler.

There's still a short initial 3MHz glitch I was unable to fully
get rid of.

Fortunately (thanks to the high MOSFET rating) everything still works.